refactor: Move agentDisplay to ui/ — shared across right-panel and sidebar
