+++ packages/client/src/ui/agentDisplay.ts
